Fix on-remote tower models
These models are stripped of scripts etc. for performance.
The downside is that breaks the prefab relation, so we have
to update them separately.

Also put back the debug_text, deactivated, so it's there in
case we need it later

Also deleted the old JP scene now that we know the new one works.